// packages/studio-ui-codegen-react/lib/amplify-ui-props/amplify-image-props.ts
export type ImageProps = {
  width?: string;
  height?: string;
  source: string;
  alt?: string;
  maxHeight?: string;
  maxWidth?: string;
  overflow?: string;
  borderRadius?: string;
  borderWidth?: string;
};

export type AmplifyImageProps = {
  htmlWidth?: string;
  htmlHeight?: string;
  source: string;
  alt?: string;
};